18-11-21 changes
In the first Intro section "Creating Digital Transformation"   ,please change the background to a solid color (Yellow)  and remove the animation.
Front color of the menu bar to a darker shade of gray/ black when we scroll down
Resize the testimonial page, there is a lot of gap on the top and button.
For the case study, Please remove the background image and replace it with a solid color preferably purple.
Provision for the images in the case study section.
Also in the case study, remove the background color of the navigation buttons instead a slightly bigger button would look better.
Please match the color of the scroll up and scroll down button (preferably both red)
In the footer, if you would put a Google Map plugin to display our office location that would be great.
Services section icons change